Refactor

+++ resources/views/relation/details.blade.php
@@ -20,7 +20,7 @@ $relation = Relation::find(Route::Input('relation_id'));
         <button type="button" class="btn btn-primary dropdown-toggle" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">Acties&nbsp;&nbsp;<span class="caret"></span></button>
         <ul class="dropdown-menu">
             <li><a href="/relation/convert?id={{ $relation->id }}">Omzetten naar particulier</a></li>
-            <li><a href="/relation/delete?id={{ $relation->id }}&csrf={{ csrf_token() }}">Verwijderen</a>
+            <li><a href="/relation/delete?id={{ $relation->id }}&csrf={{ csrf_token() }}" onclick="return confirm('Relatie verwijderen?')">Verwijderen</a>
         </li></ul>
     </div>
 </div>
